Add loginviewmodel
